Refactor App into focused modules
