+++ src/styles/appStyles.ts
@@ -1,0 +1,1480 @@
+import { StyleSheet } from 'react-native';
+
+import { tokens } from '../theme/tokens';
+
+export const styles = StyleSheet.create({
+  safeArea: {
+    backgroundColor: tokens.bg,
+    flex: 1,
+  },
+  appShell: {
+    flex: 1,
+    backgroundColor: tokens.bg,
+  },
+  screen: {
+    flex: 1,
+    backgroundColor: tokens.bg,
+  },
+  onboardingProgress: {
+    backgroundColor: tokens.bgDeep,
+    borderRadius: 8,
+    flexDirection: 'row',
+    gap: 4,
+    padding: 4,
+  },
+  onboardingProgressDot: {
+    backgroundColor: tokens.lineSoft,
+    borderRadius: 4,
+    flex: 1,
+    height: 5,
+  },
+  onboardingProgressDotActive: {
+    backgroundColor: tokens.ink,
+  },
+  onboardingThreadBlock: {
+    gap: 8,
+  },
+  suggestedReplyWrap: {
+    gap: 8,
+    marginLeft: 42,
+  },
+  setupCardGrid: {
+    gap: 10,
+  },
+  setupOptionCard: {
+    backgroundColor: tokens.surface,
+    borderColor: tokens.line,
+    borderRadius: 8,
+    borderWidth: 1,
+    gap: 4,
+    minHeight: 68,
+    paddingHorizontal: 14,
+    paddingVertical: 11,
+  },
+  setupOptionTitle: {
+    color: tokens.ink,
+    fontSize: 14,
+    fontWeight: '900',
+    letterSpacing: 0,
+  },
+  setupOptionHelper: {
+    color: tokens.accent,
+    fontSize: 11,
+    fontWeight: '900',
+    letterSpacing: 0,
+    textTransform: 'uppercase',
+  },
+  setupOptionText: {
+    color: tokens.muted,
+    fontSize: 13,
+    letterSpacing: 0,
+    lineHeight: 19,
+  },
+  topBar: {
+    alignItems: 'center',
+    borderBottomColor: tokens.lineSoft,
+    borderBottomWidth: 1,
+    flexDirection: 'row',
+    justifyContent: 'space-between',
+    paddingHorizontal: 18,
+    paddingTop: 8,
+    paddingBottom: 7,
+  },
+  topBarLeft: {
+    alignItems: 'center',
+    flexDirection: 'row',
+    gap: 10,
+  },
+  topTitle: {
+    color: tokens.ink,
+    fontSize: 14,
+    fontWeight: '800',
+    letterSpacing: 0,
+  },
+  topMeta: {
+    color: tokens.muted,
+    fontFamily: 'monospace',
+    fontSize: 10.5,
+    letterSpacing: 0,
+    marginTop: 1,
+  },
+  coachAvatarWrap: {
+    position: 'relative',
+  },
+  coachAvatar: {
+    alignItems: 'center',
+    backgroundColor: tokens.ink,
+    justifyContent: 'center',
+  },
+  coachAvatarText: {
+    color: tokens.surface,
+    fontFamily: tokens.serif,
+    fontStyle: 'italic',
+    fontWeight: '700',
+    lineHeight: 18,
+  },
+  coachOnlineDot: {
+    backgroundColor: tokens.positive,
+    borderColor: tokens.bg,
+    borderRadius: 5,
+    borderWidth: 2,
+    bottom: -1,
+    height: 9,
+    position: 'absolute',
+    right: -1,
+    width: 9,
+  },
+  feed: {
+    gap: 13,
+    padding: 16,
+    paddingBottom: 12,
+  },
+  dateDivider: {
+    alignSelf: 'center',
+    color: tokens.muted,
+    fontSize: 10,
+    fontWeight: '700',
+    letterSpacing: 2,
+    textTransform: 'uppercase',
+  },
+  coachLine: {
+    alignItems: 'flex-start',
+    flexDirection: 'row',
+    gap: 10,
+    maxWidth: '92%',
+  },
+  coachLineFirst: {
+    marginTop: 2,
+  },
+  coachSpacer: {
+    flexShrink: 0,
+    width: 32,
+  },
+  coachText: {
+    color: tokens.ink,
+    flexShrink: 1,
+    fontFamily: tokens.serif,
+    fontSize: 18,
+    fontWeight: '400',
+    letterSpacing: 0,
+    lineHeight: 25,
+    paddingHorizontal: 2,
+    paddingVertical: 3,
+  },
+  userBubble: {
+    alignSelf: 'flex-end',
+    backgroundColor: tokens.ink,
+    borderRadius: 18,
+    maxWidth: '78%',
+    paddingHorizontal: 14,
+    paddingVertical: 9,
+  },
+  userText: {
+    color: tokens.surface,
+    fontSize: 14,
+    fontWeight: '700',
+    letterSpacing: 0,
+  },
+  dataCard: {
+    backgroundColor: tokens.surface,
+    borderColor: tokens.lineSoft,
+    borderRadius: 8,
+    borderWidth: 1,
+    gap: 12,
+    padding: 14,
+  },
+  dataCardInset: {
+    marginLeft: 42,
+  },
+  dataCardWithAccent: {
+    borderLeftWidth: 2,
+  },
+  dataCardHeader: {
+    alignItems: 'center',
+    flexDirection: 'row',
+    gap: 8,
+  },
+  dataCardAccent: {
+    borderRadius: 3,
+    height: 6,
+    width: 6,
+  },
+  sectionLabel: {
+    color: tokens.muted,
+    fontSize: 10,
+    fontWeight: '800',
+    letterSpacing: 2,
+    textTransform: 'uppercase',
+  },
+  recoveryRow: {
+    alignItems: 'center',
+    flexDirection: 'row',
+    gap: 14,
+  },
+  sourceLine: {
+    alignItems: 'center',
+    flexDirection: 'row',
+    justifyContent: 'space-between',
+    gap: 10,
+  },
+  sourceLineText: {
+    color: tokens.muted,
+    flexShrink: 1,
+    fontSize: 11,
+    fontWeight: '700',
+    letterSpacing: 0,
+  },
+  recoveryCopy: {
+    flex: 1,
+    gap: 9,
+  },
+  readinessLine: {
+    alignItems: 'baseline',
+    flexDirection: 'row',
+    gap: 8,
+  },
+  readinessLabel: {
+    color: tokens.ink,
+    fontFamily: tokens.serif,
+    fontSize: 18,
+    fontStyle: 'italic',
+    fontWeight: '500',
+    letterSpacing: 0,
+  },
+  readinessMeta: {
+    color: tokens.muted,
+    fontSize: 12,
+    letterSpacing: 0,
+  },
+  ringCenter: {
+    alignItems: 'center',
+    flex: 1,
+    justifyContent: 'center',
+  },
+  ringValue: {
+    color: tokens.ink,
+    fontSize: 24,
+    fontWeight: '900',
+    letterSpacing: 0,
+  },
+  metricGrid: {
+    flexDirection: 'row',
+    flexWrap: 'wrap',
+    gap: 8,
+  },
+  metricGridFull: {
+    flexDirection: 'row',
+    flexWrap: 'wrap',
+    gap: 10,
+  },
+  smallMetric: {
+    backgroundColor: tokens.surfaceAlt,
+    borderColor: tokens.lineSoft,
+    borderRadius: 4,
+    borderWidth: 1,
+    minHeight: 62,
+    paddingHorizontal: 12,
+    paddingVertical: 10,
+    flexBasis: '47%',
+    flexGrow: 1,
+  },
+  smallMetricLabel: {
+    color: tokens.muted,
+    fontSize: 10,
+    fontWeight: '800',
+    letterSpacing: 0,
+    textTransform: 'uppercase',
+  },
+  smallMetricValue: {
+    color: tokens.ink,
+    fontFamily: tokens.serif,
+    fontSize: 22,
+    fontWeight: '400',
+    letterSpacing: 0,
+    marginTop: 4,
+  },
+  smallMetricSub: {
+    color: tokens.muted,
+    fontSize: 11,
+    letterSpacing: 0,
+    marginTop: 2,
+  },
+  planHeader: {
+    alignItems: 'center',
+    flexDirection: 'row',
+    gap: 12,
+  },
+  planIcon: {
+    alignItems: 'center',
+    backgroundColor: tokens.ink,
+    borderRadius: 99,
+    height: 38,
+    justifyContent: 'center',
+    width: 38,
+  },
+  planCopy: {
+    flex: 1,
+  },
+  planTitle: {
+    color: tokens.ink,
+    fontFamily: tokens.serif,
+    fontSize: 22,
+    fontWeight: '400',
+    letterSpacing: 0,
+    lineHeight: 25,
+  },
+  planDetail: {
+    color: tokens.muted,
+    fontSize: 13,
+    letterSpacing: 0,
+    marginTop: 2,
+    lineHeight: 18,
+  },
+  coachPlanStats: {
+    flexDirection: 'row',
+    flexWrap: 'wrap',
+    gap: 10,
+    marginTop: 7,
+  },
+  coachPlanStat: {
+    color: tokens.ink,
+    fontSize: 13,
+    fontWeight: '800',
+    letterSpacing: 0,
+    textTransform: 'capitalize',
+  },
+  effortChart: {
+    backgroundColor: tokens.surfaceAlt,
+    borderRadius: 8,
+    height: 52,
+    justifyContent: 'center',
+    paddingHorizontal: 8,
+  },
+  strainRow: {
+    alignItems: 'center',
+    flexDirection: 'row',
+    justifyContent: 'space-between',
+  },
+  strainLabel: {
+    color: tokens.muted,
+    fontSize: 12,
+    fontWeight: '700',
+    letterSpacing: 0,
+  },
+  strainValue: {
+    color: tokens.ink,
+    fontSize: 13,
+    fontWeight: '900',
+    letterSpacing: 0,
+  },
+  helpText: {
+    color: tokens.inkSoft,
+    fontSize: 13,
+    letterSpacing: 0,
+    lineHeight: 19,
+  },
+  contextRow: {
+    alignItems: 'center',
+    flexDirection: 'row',
+    gap: 10,
+    justifyContent: 'space-between',
+  },
+  contextCopy: {
+    flex: 1,
+  },
+  contextTitle: {
+    color: tokens.ink,
+    fontSize: 13,
+    fontWeight: '700',
+    letterSpacing: 0,
+  },
+  contextText: {
+    color: tokens.muted,
+    fontSize: 11,
+    letterSpacing: 0,
+    lineHeight: 16,
+    marginTop: 2,
+  },
+  segmentRow: {
+    flexDirection: 'row',
+    gap: 6,
+  },
+  segmentButton: {
+    alignItems: 'center',
+    backgroundColor: tokens.surfaceAlt,
+    borderColor: tokens.lineSoft,
+    borderRadius: 99,
+    borderWidth: 1,
+    justifyContent: 'center',
+    minHeight: 32,
+    minWidth: 42,
+    paddingHorizontal: 9,
+  },
+  segmentButtonActive: {
+    backgroundColor: tokens.ink,
+    borderColor: tokens.ink,
+  },
+  segmentButtonText: {
+    color: tokens.inkSoft,
+    fontSize: 11,
+    fontWeight: '900',
+    letterSpacing: 0,
+  },
+  segmentButtonTextActive: {
+    color: tokens.surface,
+  },
+  singleAction: {
+    marginTop: 2,
+  },
+  warningPanel: {
+    backgroundColor: '#fff8df',
+    borderColor: '#f0dc95',
+    borderRadius: 8,
+    borderWidth: 1,
+    gap: 6,
+    padding: 12,
+  },
+  warningText: {
+    color: '#6d5600',
+    fontSize: 12,
+    fontWeight: '700',
+    letterSpacing: 0,
+  },
+  chips: {
+    borderTopColor: tokens.lineSoft,
+    borderTopWidth: 1,
+    flexDirection: 'row',
+    gap: 7,
+    paddingHorizontal: 14,
+    paddingTop: 8,
+  },
+  chip: {
+    backgroundColor: tokens.surface,
+    borderColor: tokens.line,
+    borderRadius: 99,
+    borderWidth: 1,
+    paddingHorizontal: 11,
+    paddingVertical: 7,
+  },
+  chipText: {
+    color: tokens.inkSoft,
+    fontSize: 12,
+    fontWeight: '700',
+    letterSpacing: 0,
+  },
+  composer: {
+    alignItems: 'center',
+    flexDirection: 'row',
+    gap: 8,
+    borderBottomColor: tokens.lineSoft,
+    borderBottomWidth: 1,
+    paddingHorizontal: 14,
+    paddingVertical: 10,
+  },
+  composerInput: {
+    backgroundColor: tokens.surface,
+    borderColor: tokens.line,
+    borderRadius: 99,
+    borderWidth: 1,
+    color: tokens.ink,
+    flex: 1,
+    fontSize: 14,
+    height: 42,
+    justifyContent: 'center',
+    paddingHorizontal: 14,
+  },
+  composerTextInput: {
+    color: tokens.ink,
+    flex: 1,
+    fontSize: 14,
+    includeFontPadding: false,
+    height: 40,
+    letterSpacing: 0,
+    paddingVertical: 0,
+    textAlignVertical: 'center',
+  },
+  sendButton: {
+    alignItems: 'center',
+    backgroundColor: tokens.ink,
+    borderRadius: 8,
+    height: 42,
+    justifyContent: 'center',
+    width: 42,
+  },
+  pageHeader: {
+    paddingHorizontal: 20,
+    paddingTop: 12,
+    paddingBottom: 8,
+  },
+  pageEyebrow: {
+    color: tokens.muted,
+    fontSize: 12,
+    fontWeight: '800',
+    letterSpacing: 0,
+    textTransform: 'uppercase',
+  },
+  pageTitle: {
+    color: tokens.ink,
+    fontSize: 34,
+    fontWeight: '900',
+    letterSpacing: 0,
+    marginTop: 2,
+  },
+  pageContent: {
+    gap: 14,
+    padding: 18,
+    paddingBottom: 28,
+  },
+  workoutTopBar: {
+    alignItems: 'center',
+    flexDirection: 'row',
+    justifyContent: 'space-between',
+    paddingHorizontal: 18,
+    paddingTop: 12,
+    paddingBottom: 8,
+  },
+  roundIconButton: {
+    alignItems: 'center',
+    backgroundColor: tokens.surface,
+    borderColor: tokens.line,
+    borderRadius: 99,
+    borderWidth: 1,
+    height: 36,
+    justifyContent: 'center',
+    width: 36,
+  },
+  workoutContent: {
+    gap: 14,
+    paddingHorizontal: 20,
+    paddingTop: 8,
+    paddingBottom: 112,
+  },
+  workoutIntro: {
+    color: tokens.muted,
+    fontSize: 12,
+    letterSpacing: 0,
+  },
+  summaryGrid: {
+    flexDirection: 'row',
+    flexWrap: 'wrap',
+    gap: 8,
+  },
+  listCard: {
+    backgroundColor: tokens.surface,
+    borderColor: tokens.lineSoft,
+    borderRadius: 8,
+    borderWidth: 1,
+    overflow: 'hidden',
+  },
+  dayRow: {
+    alignItems: 'center',
+    borderBottomColor: tokens.lineSoft,
+    borderBottomWidth: 1,
+    flexDirection: 'row',
+    gap: 10,
+    minHeight: 68,
+    paddingHorizontal: 12,
+  },
+  dayMain: {
+    flex: 1,
+  },
+  dayTitle: {
+    color: tokens.ink,
+    fontSize: 14,
+    fontWeight: '900',
+    letterSpacing: 0,
+  },
+  dayMeta: {
+    color: tokens.muted,
+    fontSize: 12,
+    letterSpacing: 0,
+    marginTop: 2,
+  },
+  dayStats: {
+    alignItems: 'flex-end',
+    minWidth: 62,
+  },
+  dayStat: {
+    color: tokens.ink,
+    fontSize: 13,
+    fontWeight: '900',
+    letterSpacing: 0,
+  },
+  dayStatLabel: {
+    color: tokens.muted,
+    fontSize: 10,
+    fontWeight: '800',
+    letterSpacing: 0,
+    textTransform: 'uppercase',
+  },
+  workoutRow: {
+    alignItems: 'center',
+    borderBottomColor: tokens.lineSoft,
+    borderBottomWidth: 1,
+    flexDirection: 'row',
+    gap: 10,
+    minHeight: 66,
+    paddingHorizontal: 12,
+  },
+  workoutIcon: {
+    alignItems: 'center',
+    backgroundColor: tokens.surfaceAlt,
+    borderRadius: 8,
+    height: 38,
+    justifyContent: 'center',
+    width: 38,
+  },
+  workoutCopy: {
+    flex: 1,
+  },
+  workoutTitle: {
+    color: tokens.ink,
+    fontSize: 14,
+    fontWeight: '900',
+    letterSpacing: 0,
+  },
+  workoutMeta: {
+    color: tokens.muted,
+    fontSize: 12,
+    letterSpacing: 0,
+    marginTop: 2,
+  },
+  workoutWelcome: {
+    color: tokens.ink,
+    fontFamily: tokens.serif,
+    fontSize: 29,
+    fontWeight: '400',
+    letterSpacing: 0,
+    lineHeight: 34,
+  },
+  workoutSubcopy: {
+    color: tokens.muted,
+    fontSize: 14,
+    letterSpacing: 0,
+    lineHeight: 21,
+  },
+  workoutMetaRow: {
+    alignItems: 'center',
+    flexDirection: 'row',
+    gap: 8,
+  },
+  capturePill: {
+    backgroundColor: tokens.accentSoft,
+    borderRadius: 8,
+    paddingHorizontal: 10,
+    paddingVertical: 6,
+  },
+  capturePillText: {
+    color: tokens.accentDeep,
+    fontSize: 11,
+    fontWeight: '900',
+    letterSpacing: 0,
+    textTransform: 'uppercase',
+  },
+  capturePanel: {
+    backgroundColor: tokens.surfaceAlt,
+    borderRadius: 4,
+    gap: 5,
+    paddingHorizontal: 13,
+    paddingVertical: 12,
+  },
+  captureTitleRow: {
+    alignItems: 'center',
+    flexDirection: 'row',
+    gap: 8,
+  },
+  captureTitle: {
+    color: tokens.ink,
+    fontSize: 14,
+    fontWeight: '900',
+    letterSpacing: 0,
+  },
+  effortHeader: {
+    alignItems: 'center',
+    flexDirection: 'row',
+    justifyContent: 'flex-end',
+    marginTop: -28,
+  },
+  effortMeta: {
+    color: tokens.muted,
+    fontSize: 12,
+    letterSpacing: 0,
+  },
+  effortChartLarge: {
+    height: 86,
+    justifyContent: 'center',
+  },
+  durationBadge: {
+    color: tokens.ink,
+    fontSize: 12,
+    fontWeight: '900',
+    letterSpacing: 0,
+  },
+  metricTags: {
+    flexDirection: 'row',
+    flexWrap: 'wrap',
+    gap: 7,
+  },
+  metricTag: {
+    backgroundColor: tokens.surfaceAlt,
+    borderColor: tokens.lineSoft,
+    borderRadius: 8,
+    borderWidth: 1,
+    paddingHorizontal: 9,
+    paddingVertical: 6,
+  },
+  metricTagText: {
+    color: tokens.inkSoft,
+    fontSize: 11,
+    fontWeight: '800',
+    letterSpacing: 0,
+  },
+  planDayRow: {
+    alignItems: 'center',
+    borderBottomColor: tokens.lineSoft,
+    borderBottomWidth: 1,
+    flexDirection: 'row',
+    gap: 10,
+    minHeight: 68,
+    paddingHorizontal: 12,
+  },
+  planDayRowActive: {
+    backgroundColor: tokens.surfaceAlt,
+  },
+  planDayLabel: {
+    alignItems: 'center',
+    minWidth: 46,
+  },
+  planDayText: {
+    color: tokens.muted,
+    fontSize: 11,
+    fontWeight: '900',
+    letterSpacing: 0,
+    textTransform: 'uppercase',
+  },
+  coachDock: {
+    backgroundColor: tokens.bg,
+    borderTopColor: tokens.lineSoft,
+    borderTopWidth: 1,
+    gap: 10,
+    paddingHorizontal: 14,
+    paddingVertical: 10,
+  },
+  coachDockHeader: {
+    alignItems: 'center',
+    flexDirection: 'row',
+    justifyContent: 'space-between',
+  },
+  coachDockTitle: {
+    color: tokens.ink,
+    fontSize: 14,
+    fontWeight: '900',
+    letterSpacing: 0,
+  },
+  coachDockMessages: {
+    gap: 7,
+  },
+  coachDockMessagesScroll: {
+    maxHeight: 160,
+  },
+  coachDockMessage: {
+    alignSelf: 'flex-start',
+    backgroundColor: tokens.surfaceAlt,
+    borderColor: tokens.lineSoft,
+    borderRadius: 8,
+    borderWidth: 1,
+    maxWidth: '90%',
+    paddingHorizontal: 10,
+    paddingVertical: 8,
+  },
+  coachDockMessageUser: {
+    alignSelf: 'flex-end',
+    backgroundColor: tokens.ink,
+    borderColor: tokens.ink,
+  },
+  coachDockMessageText: {
+    color: tokens.inkSoft,
+    fontSize: 13,
+    letterSpacing: 0,
+    lineHeight: 18,
+  },
+  coachDockMessageTextUser: {
+    color: tokens.surface,
+    fontWeight: '800',
+  },
+  coachDockInputRow: {
+    alignItems: 'center',
+    backgroundColor: tokens.surface,
+    borderColor: tokens.line,
+    borderRadius: 99,
+    borderWidth: 1,
+    flexDirection: 'row',
+    gap: 8,
+    height: 44,
+    paddingLeft: 12,
+    paddingRight: 4,
+  },
+  coachDockInput: {
+    backgroundColor: 'transparent',
+    borderWidth: 0,
+    color: tokens.ink,
+    flex: 1,
+    fontSize: 14,
+    includeFontPadding: false,
+    height: 42,
+    lineHeight: 18,
+    paddingHorizontal: 0,
+    paddingVertical: 0,
+    textAlignVertical: 'center',
+  },
+  coachDockSend: {
+    alignItems: 'center',
+    backgroundColor: tokens.ink,
+    borderRadius: 99,
+    height: 36,
+    justifyContent: 'center',
+    width: 36,
+  },
+  coachDockCollapse: {
+    backgroundColor: tokens.surface,
+    borderColor: tokens.line,
+    borderWidth: 1,
+  },
+  coachBubbleButton: {
+    alignItems: 'center',
+    alignSelf: 'flex-end',
+    backgroundColor: tokens.ink,
+    borderRadius: 8,
+    height: 46,
+    justifyContent: 'center',
+    width: 46,
+  },
+  workoutActions: {
+    backgroundColor: tokens.bg,
+    borderTopColor: tokens.lineSoft,
+    borderTopWidth: 1,
+    flexDirection: 'row',
+    gap: 10,
+    paddingHorizontal: 20,
+    paddingVertical: 12,
+  },
+  startWatchButton: {
+    alignItems: 'center',
+    backgroundColor: tokens.ink,
+    borderRadius: 14,
+    flex: 1,
+    flexDirection: 'row',
+    gap: 8,
+    height: 50,
+    justifyContent: 'center',
+  },
+  startWatchText: {
+    color: tokens.surface,
+    fontSize: 15,
+    fontWeight: '900',
+    letterSpacing: 0,
+  },
+  scheduleButton: {
+    alignItems: 'center',
+    backgroundColor: tokens.surface,
+    borderColor: tokens.line,
+    borderRadius: 14,
+    borderWidth: 1,
+    flex: 0.75,
+    height: 50,
+    justifyContent: 'center',
+  },
+  scheduleText: {
+    color: tokens.ink,
+    fontSize: 15,
+    fontWeight: '900',
+    letterSpacing: 0,
+  },
+  historySignalContent: {
+    gap: 14,
+    paddingHorizontal: 20,
+    paddingTop: 18,
+    paddingBottom: 18,
+  },
+  historyWindow: {
+    color: tokens.muted,
+    fontSize: 13,
+    letterSpacing: 0,
+  },
+  historyTitle: {
+    color: tokens.ink,
+    fontFamily: tokens.serif,
+    fontSize: 28,
+    fontWeight: '400',
+    letterSpacing: 0,
+    lineHeight: 30,
+    marginTop: -12,
+  },
+  historyTitleItalic: {
+    fontStyle: 'italic',
+  },
+  signalTabs: {
+    flexDirection: 'row',
+    gap: 7,
+    marginTop: 4,
+  },
+  signalTab: {
+    backgroundColor: tokens.surface,
+    borderColor: tokens.line,
+    borderRadius: 99,
+    borderWidth: 1,
+    paddingHorizontal: 14,
+    paddingVertical: 8,
+  },
+  signalTabActive: {
+    backgroundColor: tokens.ink,
+    borderColor: tokens.ink,
+  },
+  signalTabText: {
+    color: tokens.muted,
+    fontSize: 12,
+    fontWeight: '800',
+    letterSpacing: 0,
+  },
+  signalTabTextActive: {
+    color: tokens.surface,
+  },
+  signalChart: {
+    backgroundColor: tokens.surface,
+    borderColor: tokens.line,
+    borderRadius: 8,
+    borderWidth: 1,
+    padding: 16,
+  },
+  signalChartHeader: {
+    alignItems: 'flex-start',
+    flexDirection: 'row',
+    justifyContent: 'space-between',
+  },
+  signalValueLine: {
+    alignItems: 'baseline',
+    flexDirection: 'row',
+    gap: 4,
+    marginTop: 2,
+  },
+  signalValue: {
+    color: tokens.ink,
+    fontFamily: tokens.serif,
+    fontSize: 32,
+    fontWeight: '400',
+    letterSpacing: 0,
+  },
+  signalUnit: {
+    color: tokens.inkSoft,
+    fontSize: 13,
+    letterSpacing: 0,
+  },
+  signalBaseline: {
+    alignItems: 'flex-end',
+  },
+  signalBaselineText: {
+    color: tokens.muted,
+    fontSize: 12,
+    letterSpacing: 0,
+  },
+  signalDelta: {
+    color: tokens.positive,
+    fontSize: 13,
+    fontWeight: '900',
+    letterSpacing: 0,
+    marginTop: 3,
+  },
+  signalDays: {
+    flexDirection: 'row',
+    justifyContent: 'space-between',
+    marginHorizontal: 1,
+    marginTop: -10,
+  },
+  signalDayText: {
+    color: tokens.inkSoft,
+    fontSize: 12,
+    letterSpacing: 0,
+  },
+  historyMiniGrid: {
+    flexDirection: 'row',
+    gap: 10,
+  },
+  historyMiniColumn: {
+    flex: 1,
+  },
+  historyMiniValue: {
+    color: tokens.ink,
+    fontFamily: tokens.serif,
+    fontSize: 25,
+    fontWeight: '400',
+    letterSpacing: 0,
+  },
+  sleepBars: {
+    alignItems: 'flex-end',
+    flexDirection: 'row',
+    gap: 5,
+    height: 38,
+    marginTop: 10,
+  },
+  sleepBar: {
+    backgroundColor: '#b6c5d3',
+    borderRadius: 2,
+    flex: 1,
+  },
+  miniSparkWrap: {
+    height: 38,
+    justifyContent: 'center',
+    marginTop: 8,
+  },
+  emptyText: {
+    color: tokens.muted,
+    fontSize: 14,
+    letterSpacing: 0,
+    padding: 14,
+  },
+  connectCard: {
+    backgroundColor: tokens.surface,
+    borderColor: tokens.lineSoft,
+    borderRadius: 8,
+    borderWidth: 1,
+    padding: 16,
+  },
+  connectIcon: {
+    alignItems: 'center',
+    backgroundColor: tokens.accentSoft,
+    borderRadius: 8,
+    height: 54,
+    justifyContent: 'center',
+    marginBottom: 14,
+    width: 54,
+  },
+  connectTitle: {
+    color: tokens.ink,
+    fontSize: 22,
+    fontWeight: '900',
+    letterSpacing: 0,
+  },
+  connectText: {
+    color: tokens.inkSoft,
+    fontSize: 13,
+    letterSpacing: 0,
+    lineHeight: 20,
+    marginTop: 6,
+  },
+  connectMeta: {
+    color: tokens.muted,
+    fontSize: 12,
+    fontWeight: '700',
+    letterSpacing: 0,
+    marginTop: 12,
+  },
+  rangeRow: {
+    backgroundColor: tokens.bgDeep,
+    borderRadius: 8,
+    flexDirection: 'row',
+    padding: 4,
+  },
+  rangeButton: {
+    alignItems: 'center',
+    borderRadius: 6,
+    flex: 1,
+    minHeight: 36,
+    justifyContent: 'center',
+  },
+  rangeButtonActive: {
+    backgroundColor: tokens.surface,
+    borderColor: tokens.line,
+    borderWidth: 1,
+  },
+  rangeButtonText: {
+    color: tokens.muted,
+    fontSize: 13,
+    fontWeight: '900',
+    letterSpacing: 0,
+  },
+  rangeButtonTextActive: {
+    color: tokens.ink,
+  },
+  actionsRow: {
+    flexDirection: 'row',
+    gap: 10,
+  },
+  button: {
+    alignItems: 'center',
+    backgroundColor: tokens.surface,
+    borderColor: tokens.line,
+    borderRadius: 8,
+    borderWidth: 1,
+    flex: 1,
+    flexDirection: 'row',
+    gap: 7,
+    justifyContent: 'center',
+    minHeight: 46,
+    paddingHorizontal: 10,
+  },
+  buttonPrimary: {
+    backgroundColor: tokens.ink,
+    borderColor: tokens.ink,
+  },
+  buttonDanger: {
+    backgroundColor: '#fff4f2',
+    borderColor: '#ffd4cc',
+  },
+  buttonText: {
+    color: tokens.ink,
+    fontSize: 14,
+    fontWeight: '900',
+    letterSpacing: 0,
+  },
+  buttonPrimaryText: {
+    color: tokens.surface,
+  },
+  buttonDangerText: {
+    color: tokens.danger,
+  },
+  disabled: {
+    opacity: 0.45,
+  },
+  pressed: {
+    opacity: 0.78,
+  },
+  scoreList: {
+    backgroundColor: tokens.surface,
+    borderColor: tokens.line,
+    borderRadius: 8,
+    borderWidth: 1,
+    overflow: 'hidden',
+  },
+  scoreRow: {
+    alignItems: 'flex-start',
+    borderBottomColor: tokens.lineSoft,
+    borderBottomWidth: 1,
+    flexDirection: 'row',
+    gap: 10,
+    minHeight: 92,
+    paddingHorizontal: 14,
+    paddingVertical: 12,
+  },
+  scoreIcon: {
+    alignItems: 'center',
+    backgroundColor: tokens.surfaceAlt,
+    borderRadius: 8,
+    height: 38,
+    justifyContent: 'center',
+    width: 38,
+  },
+  scoreCopy: {
+    flex: 1,
+    gap: 5,
+  },
+  scoreTitleLine: {
+    alignItems: 'center',
+    flexDirection: 'row',
+    gap: 8,
+    justifyContent: 'space-between',
+  },
+  scoreTitle: {
+    color: tokens.ink,
+    flex: 1,
+    fontSize: 14,
+    fontWeight: '900',
+    letterSpacing: 0,
+  },
+  scoreBadge: {
+    borderRadius: 6,
+    fontSize: 10,
+    fontWeight: '900',
+    letterSpacing: 0,
+    overflow: 'hidden',
+    paddingHorizontal: 7,
+    paddingVertical: 3,
+    textTransform: 'uppercase',
+  },
+  scoreBadge_live: {
+    backgroundColor: tokens.accentSoft,
+    color: tokens.accentDeep,
+  },
+  scoreBadge_permission: {
+    backgroundColor: '#fff4f2',
+    color: tokens.danger,
+  },
+  scoreBadge_empty: {
+    backgroundColor: '#fff7e7',
+    color: '#8a5a00',
+  },
+  scoreBadge_unchecked: {
+    backgroundColor: tokens.surfaceAlt,
+    color: tokens.muted,
+  },
+  scoreMeta: {
+    color: tokens.muted,
+    fontSize: 12,
+    letterSpacing: 0,
+  },
+  scoreTrack: {
+    backgroundColor: tokens.lineSoft,
+    borderRadius: 4,
+    height: 7,
+    overflow: 'hidden',
+  },
+  scoreFill: {
+    backgroundColor: tokens.accent,
+    borderRadius: 4,
+    height: 7,
+  },
+  scoreDetail: {
+    color: tokens.inkSoft,
+    fontSize: 11,
+    letterSpacing: 0,
+    lineHeight: 16,
+  },
+  gapList: {
+    backgroundColor: tokens.surface,
+    borderColor: tokens.line,
+    borderRadius: 8,
+    borderWidth: 1,
+    overflow: 'hidden',
+  },
+  gapRow: {
+    borderBottomColor: tokens.lineSoft,
+    borderBottomWidth: 1,
+    paddingHorizontal: 14,
+    paddingVertical: 12,
+  },
+  gapTitle: {
+    color: tokens.ink,
+    fontSize: 14,
+    fontWeight: '900',
+    letterSpacing: 0,
+  },
+  gapDetail: {
+    color: tokens.muted,
+    fontSize: 12,
+    letterSpacing: 0,
+    lineHeight: 17,
+    marginTop: 3,
+  },
+  settingsCard: {
+    backgroundColor: tokens.surface,
+    borderColor: tokens.line,
+    borderRadius: 8,
+    borderWidth: 1,
+    gap: 12,
+    padding: 14,
+  },
+  settingsHeader: {
+    alignItems: 'center',
+    flexDirection: 'row',
+    gap: 10,
+  },
+  settingsIcon: {
+    alignItems: 'center',
+    backgroundColor: tokens.accentSoft,
+    borderRadius: 8,
+    height: 36,
+    justifyContent: 'center',
+    width: 36,
+  },
+  settingsCopy: {
+    flex: 1,
+  },
+  settingsTitle: {
+    color: tokens.ink,
+    fontSize: 14,
+    fontWeight: '900',
+    letterSpacing: 0,
+  },
+  settingsMeta: {
+    color: tokens.muted,
+    fontSize: 12,
+    letterSpacing: 0,
+    marginTop: 2,
+  },
+  apiKeyInput: {
+    backgroundColor: tokens.surfaceAlt,
+    borderColor: tokens.lineSoft,
+    borderRadius: 8,
+    borderWidth: 1,
+    color: tokens.ink,
+    fontSize: 14,
+    letterSpacing: 0,
+    minHeight: 46,
+    paddingHorizontal: 12,
+  },
+  settingDivider: {
+    backgroundColor: tokens.lineSoft,
+    height: 1,
+  },
+  settingRow: {
+    alignItems: 'center',
+    flexDirection: 'row',
+    gap: 10,
+  },
+  settingSegments: {
+    backgroundColor: tokens.bgDeep,
+    borderRadius: 8,
+    flexDirection: 'row',
+    padding: 3,
+  },
+  settingSegment: {
+    alignItems: 'center',
+    borderRadius: 6,
+    justifyContent: 'center',
+    minHeight: 32,
+    minWidth: 48,
+    paddingHorizontal: 8,
+  },
+  settingSegmentActive: {
+    backgroundColor: tokens.surface,
+    borderColor: tokens.line,
+    borderWidth: 1,
+  },
+  settingSegmentText: {
+    color: tokens.muted,
+    fontSize: 12,
+    fontWeight: '900',
+    letterSpacing: 0,
+  },
+  settingSegmentTextActive: {
+    color: tokens.ink,
+  },
+  permissionList: {
+    backgroundColor: tokens.surface,
+    borderColor: tokens.line,
+    borderRadius: 8,
+    borderWidth: 1,
+    overflow: 'hidden',
+  },
+  permissionRow: {
+    alignItems: 'center',
+    borderBottomColor: tokens.lineSoft,
+    borderBottomWidth: 1,
+    flexDirection: 'row',
+    gap: 12,
+    minHeight: 62,
+    paddingHorizontal: 14,
+  },
+  permissionCopy: {
+    flex: 1,
+  },
+  permissionTitle: {
+    color: tokens.ink,
+    fontSize: 14,
+    fontWeight: '900',
+    letterSpacing: 0,
+  },
+  permissionDetail: {
+    color: tokens.muted,
+    fontSize: 12,
+    letterSpacing: 0,
+    marginTop: 2,
+  },
+  diagnosticList: {
+    backgroundColor: tokens.surface,
+    borderColor: tokens.line,
+    borderRadius: 8,
+    borderWidth: 1,
+    overflow: 'hidden',
+  },
+  diagnosticRow: {
+    alignItems: 'flex-start',
+    borderBottomColor: tokens.lineSoft,
+    borderBottomWidth: 1,
+    flexDirection: 'row',
+    gap: 10,
+    minHeight: 58,
+    paddingHorizontal: 14,
+    paddingVertical: 11,
+  },
+  diagnosticDot: {
+    backgroundColor: tokens.line,
+    borderRadius: 5,
+    height: 10,
+    marginTop: 4,
+    width: 10,
+  },
+  diagnosticDotActive: {
+    backgroundColor: tokens.positive,
+  },
+  diagnosticCopy: {
+    flex: 1,
+  },
+  diagnosticTitle: {
+    color: tokens.ink,
+    fontSize: 13,
+    fontWeight: '900',
+    letterSpacing: 0,
+  },
+  diagnosticDetail: {
+    color: tokens.muted,
+    fontSize: 11,
+    letterSpacing: 0,
+    lineHeight: 16,
+    marginTop: 2,
+  },
+  toggle: {
+    backgroundColor: tokens.line,
+    borderRadius: 8,
+    height: 24,
+    justifyContent: 'center',
+    paddingHorizontal: 2,
+    width: 40,
+  },
+  toggleActive: {
+    backgroundColor: tokens.ink,
+  },
+  toggleKnob: {
+    alignItems: 'center',
+    backgroundColor: tokens.surface,
+    borderRadius: 8,
+    height: 20,
+    justifyContent: 'center',
+    width: 20,
+  },
+  toggleKnobActive: {
+    marginLeft: 16,
+  },
+  privacyCard: {
+    alignItems: 'flex-start',
+    backgroundColor: tokens.surfaceAlt,
+    borderColor: tokens.lineSoft,
+    borderRadius: 8,
+    borderWidth: 1,
+    flexDirection: 'row',
+    gap: 10,
+    padding: 14,
+  },
+  privacyText: {
+    color: tokens.muted,
+    flex: 1,
+    fontSize: 12,
+    letterSpacing: 0,
+    lineHeight: 18,
+  },
+  tabBar: {
+    backgroundColor: tokens.surface,
+    borderTopColor: tokens.line,
+    borderTopWidth: 1,
+    flexDirection: 'row',
+    paddingHorizontal: 8,
+    paddingTop: 7,
+    paddingBottom: 5,
+  },
+  tabItem: {
+    alignItems: 'center',
+    flex: 1,
+    gap: 3,
+    minHeight: 50,
+    justifyContent: 'center',
+  },
+  tabText: {
+    color: tokens.muted,
+    fontSize: 11,
+    fontWeight: '700',
+    letterSpacing: 0,
+  },
+  tabTextActive: {
+    color: tokens.ink,
+    fontWeight: '900',
+  },
+});
