+++ src/theme/tokens.ts
@@ -1,0 +1,19 @@
+export const tokens = {
+  bg: '#fafaf8',
+  bgDeep: '#f0efea',
+  surface: '#ffffff',
+  surfaceAlt: '#f5f4f0',
+  ink: '#0a0a0a',
+  inkSoft: '#2a2a2a',
+  muted: '#6a6a6a',
+  line: '#e6e6e2',
+  lineSoft: '#f0f0ec',
+  accent: '#0a0a0a',
+  accentDeep: '#000000',
+  accentSoft: '#e8e8e4',
+  positive: '#0a7a4a',
+  warm: '#b8915a',
+  cool: '#5a7a9a',
+  danger: '#b42318',
+  serif: 'Georgia',
+};
